Refactor: exclusão do componente de selector e inserção dele diretamente no searchInput
